MAINT: move copulas to copula folder and refactor classes
