+++ statsmodels/distributions/copula/other_copulas.py
@@ -6,3 +6,33 @@ Author: Josef Perktold
 License: BSD-3
 
 """
+import numpy as np
+
+from statsmodels.distributions.copula.copulas import Copula
+
+
+class IndependentCopula(Copula):
+    """Independent copula.
+
+    .. math::
+
+        C_\theta(u,v) = uv
+
+    """
+    def __init__(self, d=2):
+        self.d = d
+        super().__init__(d=self.d)
+
+    def random(self, n=1, random_state=None):
+        rng = check_random_state(random_state)
+        x = rng.random((n, self.d))
+        return x
+
+    def pdf(self, u):
+        return np.ones((len(u), 1))
+
+    def cdf(self, u):
+        return np.prod(u, axis=1)
+
+    def plot_pdf(self, *args):
+        raise NotImplementedError("PDF is constant over the domain.")
